remove duplications for scala.examples.ml
