+++ examples/src/main/scala/org/apache/spark/examples/ml/LogisticRegressionSummaryExample.scala
@@ -18,12 +18,20 @@
 // scalastyle:off println
 package org.apache.spark.examples.ml
 
+import java.io.File
+
+import com.google.common.io.Files
+
 import org.apache.spark.{SparkConf, SparkContext}
 // $example on$
 import org.apache.spark.ml.classification.{BinaryLogisticRegressionSummary, LogisticRegression}
 // $example off$
-import org.apache.spark.sql.SQLContext
+import org.apache.spark.ml.param.ParamMap
+import org.apache.spark.mllib.linalg.Vector
+import org.apache.spark.mllib.stat.MultivariateOnlineSummarizer
 import org.apache.spark.sql.functions.max
+import org.apache.spark.sql.Row
+import org.apache.spark.sql.SQLContext
 
 object LogisticRegressionSummaryExample {
 
@@ -36,6 +44,30 @@ object LogisticRegressionSummaryExample {
     // Load training data
     val training = sqlCtx.read.format("libsvm").load("data/mllib/sample_libsvm_data.txt")
 
+    // Show statistical summary of labels.
+    val labelSummary = training.describe("label")
+    labelSummary.show()
+
+    // Convert features column to an RDD of vectors.
+    val features = training.select("features").rdd.map { case Row(v: Vector) => v }
+    val featureSummary = features.aggregate(new MultivariateOnlineSummarizer())(
+      (summary, feat) => summary.add(feat),
+      (sum1, sum2) => sum1.merge(sum2))
+    println(s"Selected features column with average values:\n ${featureSummary.mean.toString}")
+
+    // Save the records in a parquet file.
+    val tmpDir = Files.createTempDir()
+    tmpDir.deleteOnExit()
+    val outputDir = new File(tmpDir, "dataframe").toString
+    println(s"Saving to $outputDir as Parquet file.")
+    training.write.parquet(outputDir)
+
+    // Load the records back.
+    println(s"Loading Parquet file with UDT from $outputDir.")
+    val newDF = sqlCtx.read.parquet(outputDir)
+    println(s"Schema from Parquet:")
+    newDF.printSchema()
+
     val lr = new LogisticRegression()
       .setMaxIter(10)
       .setRegParam(0.3)
@@ -45,8 +77,7 @@ object LogisticRegressionSummaryExample {
     val lrModel = lr.fit(training)
 
     // $example on$
-    // Extract the summary from the returned LogisticRegressionModel instance trained in the earlier
-    // example
+    // Extract the summary from the returned LogisticRegressionModel instance.
     val trainingSummary = lrModel.summary
 
     // Obtain the objective per iteration.
@@ -54,22 +85,41 @@ object LogisticRegressionSummaryExample {
     objectiveHistory.foreach(loss => println(loss))
 
     // Obtain the metrics useful to judge performance on test data.
-    // We cast the summary to a BinaryLogisticRegressionSummary since the problem is a
-    // binary classification problem.
+    // We cast the summary to a BinaryLogisticRegressionSummary since the problem is a binary
+    // classification problem.
     val binarySummary = trainingSummary.asInstanceOf[BinaryLogisticRegressionSummary]
 
-    // Obtain the receiver-operating characteristic as a dataframe and areaUnderROC.
+    // Obtain the receiver-operating characteristic as a DataFrame and areaUnderROC.
     val roc = binarySummary.roc
     roc.show()
     println(binarySummary.areaUnderROC)
 
-    // Set the model threshold to maximize F-Measure
+    // Set the model threshold to maximize F-Measure.
     val fMeasure = binarySummary.fMeasureByThreshold
     val maxFMeasure = fMeasure.select(max("F-Measure")).head().getDouble(0)
     val bestThreshold = fMeasure.where($"F-Measure" === maxFMeasure)
       .select("threshold").head().getDouble(0)
     lrModel.setThreshold(bestThreshold)
     // $example off$
+
+    // Use extra params
+
+    // We may alternatively specify parameters using a ParamMap, which supports several methods for
+    // specifying parameters.
+    val paramMap = ParamMap(lr.maxIter -> 20)
+    // Specify 1 Param.  This overwrites the original maxIter.
+    paramMap.put(lr.maxIter, 30)
+    // Specify multiple Params.
+    paramMap.put(lr.regParam -> 0.1, lr.thresholds -> Array(0.45, 0.55))
+
+    // One can also combine ParamMaps.
+    val paramMap2 = ParamMap(lr.probabilityCol -> "myProbability")  // Change output column name.
+    val paramMapCombined = paramMap ++ paramMap2
+
+    // Now learn a new model using the paramMapCombined parameters.
+    // paramMapCombined overrides all parameters set earlier via lr.set* methods.
+    val model2 = lr.fit(training.toDF(), paramMapCombined)
+    println("Model 2 was fit using parameters: " + model2.parent.extractParamMap())
 
     sc.stop()
   }
